Finished A4Q4

+++ src/A4Qn/A4Q4.java
@@ -17,36 +17,36 @@ public class A4Q4 {
      */
     public static void main(String[] args) {
         // TODO code application logic here
-        
-        Scanner input = new Scanner (System.in);
-        
+
+        Scanner input = new Scanner(System.in);
+
         System.out.println("How much does food for prom cost?");
-        int food = input.nextInt();
-        
+        float food = input.nextInt();
+
         System.out.println("How much does the DJ cost?");
-        int DJ = input.nextInt();
-        
+        float DJ = input.nextInt();
+
         System.out.println("How much does it cost to rent the hall?");
-        int Hall = input.nextInt();
-        
+        float Hall = input.nextInt();
+
         System.out.println("How much do the decorations cost?");
-        int Dec = input.nextInt();
-        
+        float Dec = input.nextInt();
+
         System.out.println("How much does it cost for staff?");
-        int Staff = input.nextInt();
-        
+        float Staff = input.nextInt();
+
         System.out.println("How much do the miscellaneous charges cost?");
-        int Misc = input.nextInt();
-        
-        double Total = (food + DJ + Hall + Dec + Staff + Misc)/35.0;
+        float Misc = input.nextInt();
+
+        double Total = (food + DJ + Hall + Dec + Staff + Misc) / 35.0;
         double Total1 = Math.ceil(Total);
-        
-        System.out.println("The total cost is " + (food + DJ + Hall + Dec + Staff + Misc) + ". You will need to sell " + Total1 + " tickets to break even." );
-                
-                
-        
-        
-        
-        
+
+        System.out.println("The total cost is " + (food + DJ + Hall + Dec + Staff + Misc) + ". You will need to sell " + Total1 + " tickets to break even.");
+
+
+
+
+
+
     }
 }
